More styling
Mehir helped me with the styling
