+++ src/components/Sidebar/Sidebar.css
@@ -1,17 +1,18 @@
 .armada-icon {
   height: 48px;
-  width: 48x;
+  width: 48px;
   padding-bottom: 50px;
 }
 
 .sidebarFlex{
   display: flex;
   flex-direction: column;
-  justify-content: space-between;
+  align-items: center;
   gap: 10px;
-  height: 100%;
+  height: 100vh;
   background-color: white;
-  padding-right: 20px;  
+  padding-right: 20px; 
+  border: solid black;
 } 
 /*
   Help me center this shit 
@@ -22,13 +23,29 @@
   flex-direction: column;
   justify-content: center;
   gap: 10px;
-  height: 100%;
   background-color: white;
-  border:1px solid black;
+  border:1px solid rgb(0, 0, 0);
+}
+
+.button-link{
+  height: 172px;
+  width: 172px;
+}
+
+.button-active{
+  width: 100px;
+  height: 100px;
+  background-color: green
 }
 
 .sidebar-LinkButton{
- 
+  border: 1px solid black;
+  
+}
+
+.navbar-icon{
+  height: 24px;
+  width: 24px;
 }
